requirements.txt
asgiref==3.5.1
autopep8==1.6.0
backports.zoneinfo==0.2.1
Django==4.0.4
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
Pillow==9.1.0
pycodestyle==2.8.0
PyJWT==2.4.0
pytz==2022.1
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.1
gunicorn==20.1.0
psycopg2==2.9.3